# catalog/src/main/resources/api/openapi.yml
openapi: 3.0.3
info:
  title: Library Catalog
  description: Provides an API to access the catalog of a library.
  contact:
    email: acarbenay@adhuc.fr
  version: 1.0.0
servers:
  - url: https://catalog.library.adhuc.org/api
tags:
  - name: Books
    description: Access the books in the catalog of the library.
  - name: Authors
    description: Access the authors of the books in the catalog of the library.
paths:
  /v1/catalog:
    get:
      tags:
        - Books
      summary: Gets a paginated list of books based on filters.
      description: The catalog can be sorted and filtered in many different ways.
      operationId: getCatalog
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '206':
          $ref: '#/components/responses/Catalog'
  /v1/books/{id}:
    get:
      tags:
        - Books
      summary: Gets the details of a book based on its identity.
      description: |
        The details of a book contain all its information and the relation to
        its authors.
      operationId: getBook
      parameters:
        - $ref: '#/components/parameters/BookId'
      responses:
        '200':
          $ref: '#/components/responses/BookDetails'
        '404':
          $ref: '#/components/responses/UnknownBook'
  /v1/authors/{id}:
    get:
      tags:
        - Authors
      summary: Gets the details of an author based on its identity.
      description: |
        The details of an author contain all its information and the relation to
        its written books.
      operationId: getAuthor
      parameters:
        - $ref: '#/components/parameters/AuthorId'
      responses:
        '200':
          $ref: '#/components/responses/AuthorDetails'
        '404':
          $ref: '#/components/responses/UnknownAuthor'
components:
  parameters:
    BookId:
      name: id
      description: Book identity.
      in: path
      required: true
      schema:
        type: string
        format: uuid
        example: baa19c4e-a347-45bd-8caf-e150aa8203e9
    AuthorId:
      name: id
      description: Author identity.
      in: path
      required: true
      schema:
        type: string
        format: uuid
        example: 99287cef-2c8c-4a4d-a82e-f1a8452dcfe2
    Page:
      name: page
      description: |
        Page number to retrieve from the collection, starting from 0. This
        parameter must be used in combination with the *size* parameter to
        specify the number of elements in a page.
      in: query
      required: false
      schema:
        type: number
        minimum: 0
        default: 0
        example: 2
    PageSize:
      name: size
      description: |
        Size of the page to retrieve from the collection. This parameter must be
        used in combination with the *page* parameter to specify the current page.
      in: query
      required: false
      schema:
        type: number
        minimum: 1
        default: 50
        example: 50
  responses:
    Catalog:
      description: Books returned from catalog.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Catalog'
        application/hal+json:
          schema:
            $ref: '#/components/schemas/Catalog'
    BookDetails:
      description: A book details.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BookDetails'
        application/hal+json:
          schema:
            $ref: '#/components/schemas/BookDetails'
    UnknownBook:
      description: No book corresponds to the specified identity.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unknown:
              $ref: '#/components/examples/UnknownBook'
        application/hal+json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unknown:
              $ref: '#/components/examples/UnknownBook'
    AuthorDetails:
      description: An author details.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthorDetails'
        application/hal+json:
          schema:
            $ref: '#/components/schemas/AuthorDetails'
    UnknownAuthor:
      description: No author corresponds to the specified identity.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unknown:
              $ref: '#/components/examples/UnknownAuthor'
        application/hal+json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unknown:
              $ref: '#/components/examples/UnknownAuthor'
  schemas:
    Catalog:
      type: object
      required:
        - page
        - '_links'
      properties:
        page:
          $ref: '#/components/schemas/Page'
        '_embedded':
          type: object
          readOnly: true
          required:
            - books
            - authors
          properties:
            books:
              description: The books collection, containing basic information for each book.
              type: array
              items:
                $ref: '#/components/schemas/Book'
            authors:
              description: The authors related to the books in the collection.
              type: array
              items:
                $ref: '#/components/schemas/Author'
        '_links':
          type: object
          readOnly: true
          required:
            - self
          properties:
            self:
              description: Link to this catalog page.
              type: object
              required:
                - href
              properties:
                href:
                  type: string
                  example: 'https://catalog.library.adhuc.org/api/v1/catalog?page=2&size=50'
            template:
              description: Link to the template of catalog page.
              type: object
              required:
                - href
                - templated
              properties:
                href:
                  type: string
                  example: 'https://catalog.library.adhuc.org/api/v1/catalog{?page,size}'
                templated:
                  type: boolean
                  enum:
                    - true
            prev:
              description: Link to the previous catalog page.
              type: object
              required:
                - href
              properties:
                href:
                  type: string
                  example: 'https://catalog.library.adhuc.org/api/v1/catalog?page=1&size=50'
            next:
              description: Link to the next catalog page.
              type: object
              required:
                - href
              properties:
                href:
                  type: string
                  example: 'https://catalog.library.adhuc.org/api/v1/catalog?page=3&size=50'
            first:
              description: Link to the first catalog page.
              type: object
              required:
                - href
              properties:
                href:
                  type: string
                  example: 'https://catalog.library.adhuc.org/api/v1/catalog?page=0&size=50'
            last:
              description: Link to the last catalog page.
              type: object
              required:
                - href
              properties:
                href:
                  type: string
                  example: 'https://catalog.library.adhuc.org/api/v1/catalog?page=8&size=50'
    Book:
      type: object
      required:
        - id
        - isbn
        - title
        - authors
        - language
        - summary
        - '_links'
      properties:
        id:
          $ref: '#/components/schemas/BookId'
        isbn:
          $ref: '#/components/schemas/ISBN'
        title:
          $ref: '#/components/schemas/Title'
        authors:
          $ref: '#/components/schemas/BookAuthors'
        language:
          $ref: '#/components/schemas/Language'
        summary:
          $ref: '#/components/schemas/Summary'
        '_links':
          type: object
          readOnly: true
          required:
            - self
          properties:
            self:
              description: Link to this book. Can be used to see the details of this book.
              type: object
              required:
                - href
              properties:
                href:
                  type: string
                  example: 'https://catalog.library.adhuc.org/api/v1/books/baa19c4e-a347-45bd-8caf-e150aa8203e9'
    BookDetails:
      type: object
      required:
        - id
        - isbn
        - title
        - publication_date
        - language
        - summary
        - '_links'
      properties:
        id:
          $ref: '#/components/schemas/BookId'
        isbn:
          $ref: '#/components/schemas/ISBN'
        title:
          $ref: '#/components/schemas/Title'
        publication_date:
          $ref: '#/components/schemas/PublicationDate'
        language:
          $ref: '#/components/schemas/Language'
        summary:
          $ref: '#/components/schemas/Summary'
        '_links':
          type: object
          readOnly: true
          required:
            - self
          properties:
            self:
              description: Link to this book. Can be used to see the details of this book.
              type: object
              required:
                - href
              properties:
                href:
                  type: string
                  example: 'https://catalog.library.adhuc.org/api/v1/books/baa19c4e-a347-45bd-8caf-e150aa8203e9'
        '_embedded':
          type: object
          readOnly: true
          required:
            - authors
          properties:
            authors:
              description: The authors related to the book.
              type: array
              items:
                $ref: '#/components/schemas/Author'
    BookId:
      type: string
      format: uuid
      example: baa19c4e-a347-45bd-8caf-e150aa8203e9
    ISBN:
      type: string
      format: isbn
      example: 9782081275232
    Title:
      type: string
      example: Du contrat social
    PublicationDate:
      type: string
      description: |
        The publication date can be either an exact date, in 'yyy-MM-dd' format,
        or a year, either negative (BC) or positive (AD).
      example: 1762
    BookAuthors:
      type: array
      items:
        $ref: '#/components/schemas/AuthorId'
    Author:
      type: object
      required:
        - id
        - name
        - '_links'
      properties:
        id:
          $ref: '#/components/schemas/AuthorId'
        name:
          $ref: '#/components/schemas/AuthorName'
        '_links':
          type: object
          readOnly: true
          required:
            - self
          properties:
            self:
              description: Link to this author. Can be used to see the detail of this author.
              type: object
              required:
                - href
              properties:
                href:
                  type: string
                  example: 'https://catalog.library.adhuc.org/api/v1/authors/99287cef-2c8c-4a4d-a82e-f1a8452dcfe2'
    AuthorDetails:
      type: object
      required:
        - id
        - name
        - date_of_birth
        - '_links'
      properties:
        id:
          $ref: '#/components/schemas/AuthorId'
        name:
          $ref: '#/components/schemas/AuthorName'
        date_of_birth:
          $ref: '#/components/schemas/DateOfBirth'
        date_of_death:
          $ref: '#/components/schemas/DateOfDeath'
        '_links':
          type: object
          readOnly: true
          required:
            - self
          properties:
            self:
              description: Link to this author. Can be used to see the detail of this author.
              type: object
              required:
                - href
              properties:
                href:
                  type: string
                  example: 'https://catalog.library.adhuc.org/api/v1/authors/99287cef-2c8c-4a4d-a82e-f1a8452dcfe2'
        '_embedded':
          type: object
          readOnly: true
          required:
            - notable_books
          properties:
            books:
              description: The author's books, containing basic information for each book.
              type: array
              items:
                $ref: '#/components/schemas/Book'
    AuthorId:
      type: string
      format: uuid
      example: 99287cef-2c8c-4a4d-a82e-f1a8452dcfe2
    AuthorName:
      type: string
      example: Jean-Jacques Rousseau
    DateOfBirth:
      type: string
      format: date
      example: 1712-06-28
    DateOfDeath:
      type: string
      format: date
      example: 1778-07-02
    NotableBooks:
      type: array
      items:
        $ref: '#/components/schemas/BookId'
    Language:
      type: string
      example: French
    Summary:
      type: string
      example: >
        Paru en 1762, le Contrat social, en affirmant le principe de souveraineté du peuple,
        a constitué un tournant décisif pour la modernité et s'est imposé comme un des textes
        majeurs de la philosophie politique. Il a aussi acquis le statut de monument, plus
        célèbre que connu, plus révéré - ou honni - qu'interrogé. Retrouver, dans les formules
        fameuses et les pages d'anthologie, le mouvement de la réflexion et les questions vives
        qui nourrissent une œuvre beaucoup plus problématique qu'affirmative, c'est découvrir
        une pensée qui se tient au plus près des préoccupations d'aujourd'hui : comment intégrer
        les intérêts de tous dans la détermination de l'intérêt commun ? Comment lutter contre
        la pente de tout gouvernement à déposséder les citoyens de la souveraineté ? Comment
        former en chacun ce sentiment d'obligation sans lequel le lien social se défait ?
    Page:
      type: object
      readOnly: true
      required:
        - size
        - total_elements
        - total_pages
        - number
      properties:
        size:
          description: The requested page size. The number of elements in the current page is between 0 and this size.
          type: number
          example: 50
          minimum: 0
        total_elements:
          description: The number of elements in all pages.
          type: number
          example: 443
          minimum: 0
        total_pages:
          description: The number of pages containing all the elements.
          type: number
          example: 9
          minimum: 0
        number:
          description: The current page, starting at 0.
          type: number
          example: 2
          minimum: 0
    Error:
      type: object
      required:
        - timestamp
        - status
        - error
        - description
      properties:
        timestamp:
          type: string
          description: The server timestamp when generating the error.
          format: date-time
        status:
          type: integer
          description: The HTTP status of the response.
        error:
          type: string
          description: The error general code.
        description:
          type: string
          description: The error specific description to help understanding and resolution.
        details:
          type: array
          description: The error details, providing information about each problem.
          items:
            type: object
            required:
              - reason
            properties:
              reason:
                type: string
                description: The problem reason used to determine why the server returned an error.
              pointer:
                type: string
                description: |
                  A JSON pointer (https://datatracker.ietf.org/doc/html/rfc6901) to the associated entity in the request
                  document (e.g. "/" for a primary object, "/name" for a specific attribute named "name").
              parameter:
                type: string
                description: A string indicating which URI query parameter caused the error.
  examples:
    UnknownBook:
      value:
        timestamp: '2024-01-01T12:34:56'
        status: 404
        error: ENTITY_NOT_FOUND
        description: No book exists with id 'e5c0f8ae-2092-4acb-92e4-0088870ddf43'
      summary: Error when no book corresponds to the specified identity.
    UnknownAuthor:
      value:
        timestamp: '2024-01-01T12:34:56'
        status: 404
        error: ENTITY_NOT_FOUND
        description: No author exists with id '96c7774f-dd6b-43d4-bd0e-6a7d1aad09a4'
      summary: Error when no author corresponds to the specified identity.